requirements.txt
streamlit
torch
torchvision
numpy
Pillow
psycopg2-binary
streamlit-drawable-canvas